Update 02.extra-4.js
I decided to put this note in the final instead of the exercise notes to avoid giving away the answer accidentally. Anyone who understands the JS prerequisites should be able to get past it on their own anyway, and by the time the learner has come to the final they're expecting to see the answer. So putting it here keeps the learning path the same and only derails the part of the process where they jump over to discord to ask what it means.