Style: "co-worker" -> "coworker"
Peter the Product Manager already introduced Kellie as a "coworker,"
which is how it is spelled in American English, according to the Oxford
English Dictionary. The OED says "co-worker" is an EN-GB thing.

I dove into this one a little bit and while the spelling is disputed by
AP Stylebook, I treat OED as the canonical source of English, and since
the rest of this material is in EN-US, let's go with the American
English version.

For whatever it's worth, Merriam-Webster doesn't recognize "co-worker"
at all.

https://www.grammarly.com/blog/commonly-confused-words/co-worker-vs-coworker/